Make targetBoxYPercentage validation use floats
Make targetBoxYPercentage validation use floating
point numbers

+++ controllers/api/v1/playController.js
@@ -14,7 +14,7 @@ const changeGameStatePutValidator = [
         .notEmpty()
         .isFloat({ max: 100 }),
     body("targetBoxYPercentage")
-        .toInt()
+        .toFloat()
         .notEmpty()
 ]
